+++ src/index.js
@@ -9,14 +9,16 @@
 const mustache = require("mustache");
 const prettier = require("prettier");
 const _ = require("lodash");
+const { resolve } = require("path");
 const { parseSchemas } = require("./schema");
 const { parseRoutes, groupRoutes } = require("./routes");
 const { createApiConfig } = require("./apiConfig");
 const { getModelType } = require("./modelTypes");
 const { getSwaggerObject, fixSwaggerScheme } = require("./swagger");
 const { createComponentsMap, filterComponentsMap } = require("./components");
-const { getTemplate, createFile, pathIsExist } = require("./files");
+const { createFile, pathIsExist } = require("./files");
 const { addToConfig, config } = require("./config");
+const { getTemplates } = require("./templates");
 
 mustache.escape = (value) => value;
 
@@ -33,10 +35,12 @@ module.exports = {
     output,
     url,
     name,
+    templates = resolve(__dirname, config.templates),
     generateResponses = config.generateResponses,
     defaultResponseAsSuccess = config.defaultResponseAsSuccess,
     generateRouteTypes = config.generateRouteTypes,
     generateClient = config.generateClient,
+    generateUnionEnums = config.generateUnionEnums,
   }) =>
     new Promise((resolve, reject) => {
       addToConfig({
@@ -44,9 +48,13 @@ module.exports = {
         generateRouteTypes,
         generateClient,
         generateResponses,
+        templates,
+        generateUnionEnums,
       });
       getSwaggerObject(input, url)
         .then(({ usageSchema, originalSchema }) => {
+          const { apiTemplate, clientTemplate, routeTypesTemplate } = getTemplates();
+
           console.log("☄️  start generating your typescript api");
 
           fixSwaggerScheme(usageSchema, originalSchema);
@@ -57,10 +65,6 @@ module.exports = {
           });
 
           const { info, paths, servers, components } = usageSchema;
-
-          const apiTemplate = getTemplate("api");
-          const clientTemplate = getTemplate("client");
-          const routeTypesTemplate = getTemplate("route-types");
 
           const componentsMap = createComponentsMap(components);
           const schemasMap = filterComponentsMap(componentsMap, "schemas");